Factorize and cleanup panner code.

+++ source/common/spat.h
@@ -4,6 +4,6 @@
  ************************************************************************/
 
 // spatialisation
-float spat(float x, float d, int n);
-float spat2(float x, float d);
-void  panner(double* out, int n, double teta, double d);
+double spat(double x, double d, int n);
+double spat2(double x, double d);
+void   panner(double* out, int n, double teta, double d);
